Hobby-Policy anwenden

+++ app/Http/Controllers/HobbyController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Session;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyController extends Controller
 {
@@ -118,6 +119,13 @@ class HobbyController extends Controller
      */
     public function edit(Hobby $hobby)
     {
+
+        if (auth()->guest()) {
+            abort(403);
+        }
+
+        abort_unless($hobby->user_id === auth()->id() || auth()->user()->rolle === 'admin', 403);
+
         return view('hobby.edit')->with('hobby', $hobby);
     }
 
@@ -130,6 +138,9 @@ class HobbyController extends Controller
      */
     public function update(Request $request, Hobby $hobby)
     {
+
+        abort_unless(Gate::allows('update', $hobby), 403);
+
         $request->validate(
             [
                 'name' => 'required|min:3',
@@ -160,6 +171,14 @@ class HobbyController extends Controller
      */
     public function destroy(Hobby $hobby)
     {
+
+        if (auth()->guest()) {
+            abort(403);
+        }
+
+        abort_unless(Gate::allows('delete', $hobby), 403);
+
+
         $old_name = $hobby->name;
         $hobby->delete();
         return back()->with([
